+++ src/net/sf/mzmine/modules/alignment/fragment/AlignFragmentTask.java
@@ -46,12 +46,12 @@ class AlignFragmentTask implements Task {
 	private int processedRows, totalRows;
 
 	private String peakListName;
-	private float mzTolerance, mzWeight;
+	private double mzTolerance, mzWeight;
 	private boolean rtToleranceUseAbs;
-	private float rtToleranceValueAbs, rtToleranceValuePercent;
-	private float rtWeight;
+	private double rtToleranceValueAbs, rtToleranceValuePercent;
+	private double rtWeight;
 	private boolean sameIDRequired;
-	private float sameIDWeight;
+	private double sameIDWeight;
 	
 	private FragmentAligner fragmentAligner;
 
@@ -69,23 +69,23 @@ class AlignFragmentTask implements Task {
 		this.fragmentAligner = fragmentAligner;
 
 		// Get parameter values for easier use
-		mzTolerance = (Float) parameters
+		mzTolerance = (Double) parameters
 				.getParameterValue(FragmentAlignerParameters.MZTolerance);
-		mzWeight = (Float) parameters
+		mzWeight = (Double) parameters
 				.getParameterValue(FragmentAlignerParameters.MZWeight);
 
 		rtToleranceUseAbs = (parameters
 				.getParameterValue(FragmentAlignerParameters.RTToleranceType) == FragmentAlignerParameters.RTToleranceTypeAbsolute);
-		rtToleranceValueAbs = (Float) parameters
+		rtToleranceValueAbs = (Double) parameters
 				.getParameterValue(FragmentAlignerParameters.RTToleranceValueAbs);
-		rtToleranceValuePercent = (Float) parameters
+		rtToleranceValuePercent = (Double) parameters
 				.getParameterValue(FragmentAlignerParameters.RTToleranceValuePercent);
-		rtWeight = (Float) parameters
+		rtWeight = (Double) parameters
 				.getParameterValue(FragmentAlignerParameters.RTWeight);
 
 		sameIDRequired = (Boolean) parameters
 				.getParameterValue(FragmentAlignerParameters.SameIDRequired);
-		sameIDWeight = (Float) parameters
+		sameIDWeight = (Double) parameters
 				.getParameterValue(FragmentAlignerParameters.SameIDWeight);
 
 	}
@@ -100,10 +100,10 @@ class AlignFragmentTask implements Task {
 	/**
 	 * @see net.sf.mzmine.taskcontrol.Task#getFinishedPercentage()
 	 */
-	public float getFinishedPercentage() {
+	public double getFinishedPercentage() {
 		if (totalRows == 0)
 			return 0f;
-		return (float) processedRows / (float) totalRows;
+		return (double) processedRows / (double) totalRows;
 	}
 
 	/**
@@ -180,10 +180,10 @@ class AlignFragmentTask implements Task {
 					return;
 
 				// Calculate limits for a row with which the row can be aligned
-				float mzMin = row.getAverageMZ() - mzTolerance;
-				float mzMax = row.getAverageMZ() + mzTolerance;
-				float rtMin, rtMax;
-				float rtToleranceValue = 0.0f;
+				double mzMin = row.getAverageMZ() - mzTolerance;
+				double mzMax = row.getAverageMZ() + mzTolerance;
+				double rtMin, rtMax;
+				double rtToleranceValue = 0.0f;
 				if (rtToleranceUseAbs) {
 					rtToleranceValue = rtToleranceValueAbs;
 					rtMin = row.getAverageRT() - rtToleranceValue;
